+++ Horiba.Sdk.Examples/Examples/CCD/CcdAbortExample.cs
@@ -3,7 +3,7 @@ using Horiba.Sdk.Devices;
 using Horiba.Sdk.Enums;
 using Newtonsoft.Json.Linq;
 
-namespace Horiba.Sdk.Examples.CcdExamples
+namespace Horiba.Sdk.Examples.Ccd
 {
     public class CcdAbortExample
     {
