+++ Horiba.Sdk.Examples/Examples/CCD/CcdDarkCountSubtractionExample.cs
@@ -7,7 +7,7 @@ using Horiba.Sdk.Devices;
 using Horiba.Sdk.Enums;
 using Serilog;
 
-namespace Horiba.Sdk.Examples.CcdExamples
+namespace Horiba.Sdk.Examples.Ccd
 {
     public class CcdDarkCountSubtractionExample
     {
