+++ Horiba.Sdk.Examples/Examples/CCD/CcdParallelSpeedExample.cs
@@ -3,7 +3,7 @@ using Horiba.Sdk.Devices;
 using Horiba.Sdk.Enums;
 using Serilog;
 
-namespace Horiba.Sdk.Examples.CcdExamples
+namespace Horiba.Sdk.Examples.Ccd
 {
     public class CcdParallelSpeedExample
     {
